RangeControllers now auto-configure!!11one
RangeControllers now search their
child hierarchy for all RangeTargets and associates with them automatically. To place new targets, simply copy and paste the TargetGroup.

RangeTargets now auto-configer all their Hints - all that is needed is for a Hint component to exist on them, and they must be a hierarchical sibling.

RangeControllers also now find all ConsoleButtons in its hierarchy and configures them - as well as triggers all of them if any one is pushed. This allows good behaviour if you have multiple buttons in one Range, which is now simply a copy paste job.

+ TargetGroups

TargetGroups is simply a logical container for a Target and any number of Hints related to that target. To create new targets and hints, all you need to do is make sure they exist in a TargetGroup below the Controller in the hierarchy, and they'll be picked up.