Various tweaks
Playing around with the origin point of projectile weapon, as well as changed the GRAPHICAL origin point of the rail gun, such that the beam is actually visible after being fired. Previously, the origin of the line was centered on the player camera, and was hence invisible until you moved a few cm.

The new railgun origin is purely graphical - the actual raycast for hit detection is still cast from the center of the player camera.