Added PlayerPref  string "playerName" - required
The game forces you to enter a playerName if none exists, and checks to make sure the entered string is not null or blank. Additionally allows you to hit F10 to edit it, and if it already exists, defaults to showing said string in the editbox when you hit F10.